+++ app/src/main/java/com/codetutor/countryinfoapp/database/appdb/DatabaseProvider.kt
@@ -1,0 +1,7 @@
+package com.codetutor.countryinfoapp.database.appdb
+
+import com.codetutor.countryinfoapp.database.dao.CountryDao
+
+interface DatabaseProvider {
+    fun countryDao(): CountryDao
+}
