+++ app/src/main/java/com/codetutor/countryinfoapp/database/dao/CountryDao.kt
@@ -1,6 +1,5 @@
-package com.codetutor.countryinfoapp.database
+package com.codetutor.countryinfoapp.database.dao
 
-import CountryEntity
 import androidx.room.Dao
 import androidx.room.Delete
 import androidx.room.Insert
@@ -10,22 +9,22 @@ import androidx.room.Update
 import com.codetutor.countryinfoapp.data.Country
 
 @Dao
-interface CountryDao {
+interface CountryDao: ICountryDao {
     @Query("SELECT * FROM Country")
-    suspend fun getAllCountries(): List<Country>
+    override suspend fun getAllCountries(): List<Country>
 
     @Query("SELECT * FROM Country WHERE continents LIKE :continent")
-    suspend fun getCountriesByContinent(continent: String): List<Country>
+    override suspend fun getCountriesByContinent(continent: String): List<Country>
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insertAll(countries: List<Country>)
+    override suspend fun insertAll(countries: List<Country>)
 
     @Delete
-    suspend fun delete(country: Country)
+    override suspend fun delete(country: Country)
 
     @Query("Update Country set capital = :capital where id = :id")
-    suspend fun updateCapital(capital: List<String>, id: Int): Int
+    override suspend fun updateCapital(capital: List<String>, id: Int): Int
 
     @Update(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun updateCountry(country: Country): Int
+    override suspend fun updateCountry(country: Country): Int
 }
